refactor: RenderToDom / addButtonFunctionality
Refactor RenderToDom as the original was calling addButtonFunctionality
each time a column was clicked - thereby adding an extra eventListener
to each column. RenderToDom now includes a check for "missed" shots.
If a target is missed a new span will be added to DOM denoting the missed
shot.

addButtonFunctionality now checks to see if a the event target has already
been clicked. If so, it returns a message to the user without re-rendering
the gameboards. If the event target has NOT been clicked an attack is made
to the board and RenderToDom function is called, thereby re-rendering
boards.