feat: add attackPlayer fn
Add attackPlayer function so that when the player attacks the computer,
the computer will reciprocate with a response attack after a 1200ms
setTimeout delay.

Also, in the attack call to the gameController, pass in the attacker
along with the oppenent. This allows more responsive messaging in the
banner i.e. "Sam has sunk the patrol vessel"