fix: remove extra div and boat image classes
Remove boat name classes from boat-container divs as boat-image classes
were handling the sizing just fine with the exception of submarine. Moved
the 'submarine' class directly to the image to further size its width.
Remove the div surrounding #start-screen-gameboard as it was no longer
needed since 'Drag and drop to place' was moved to boat-placement-container