feat: add new class/data-attr to boat-containers
Add 'hover-effect' class and data-boat-container to all boat-container
divs. The hover effect was originally a pseudo class for all boat images.
Once user places a boat on the board, the pseudo class was still in
effect causing a bug with the newly desired hover effect of showing a
'no-click' cursor while hovering over an already-placed boat. By turning
the hover-effect into a class, the cursor can be altered after boat
placement.

Add data attribute to make the entire boat-container clickable instead
of just the image.