fix: add guard clause - handleHoverRemoveHighlight
Add class check to handleHoverRemoveHighlight to handle edge case. This
ensures that if column containing the 'remove-pointer' class is hovered
over, the mouseout listener will not take affect. Though 'remove-pointer'
was preventing clicks and further placement successfully, the mouseout
event was still taking effect as it is placed on the board itself and
not the column