fix: add 'remove-pointer' check to listener fns
Add 'remove-pointer' check to all listener functions as the 'remove-pointer'
class is reserved for columns that already contain a boat. This ensures
that adding highlights/removing highlights don't overlap the boats that
have already been placed. It ensures that these actions will only occur
IF the finalColumns arr does -not- contain a column with 'remove-pointer'
class