fix: add clickedAvailableSquares check to click
Add 'clickedAvailableSquares' check to the click listener. While the
'hasRemovePointer' checks within the click listener's helper function
was preventing the boat from being placed, a 'click' event was still
registering, thereby causing the current boat container to disable itself
and go dark. The 'clickedAvailableSquares' is a guard clause that prevents
this occurence.