fix: add a check for 'undefined' in click listener
Add a check for 'undefined' to the event listener. As with the previous
issue, if user attempted to use the click to place a boat wherein the length
exceeded the boundary of the board, the boat container would go dark and
become unclickable. Even though the boat wouldn't be placed in the backend
onto the array, the 'disable' class would kick in to the boat container.
This is solved with the 'undefined' check