fix: add check for alreadySelectedBoat
Add a check for alreadySelectedBoat to handleBoatContainerClick fn.
Previously, if user selected a boat then decided that they wanted to
change the selected boat before placement, both the original boat and
new boat would remain selected. Now, the 'selected' class is removed
from the alreadySelectedBoat and added to the new boat.