fix: move helper in handleBoatContainerClick
Move checkForAlreadySelectedBoat in handleBoatContainerClick fn. This
check for the already selected boat was happening whether the click was
on a new boat OR the axis button. This created issues with the axis button
where the boat would lose its 'selected' class if the axis button was
toggled before the boat placement.